pyproject.toml
[project]
name = "immich-smart-album"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32.5",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "rich>=14.3.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.6",
    "ty>=0.0.23",
]
